Revert "feat(context filter): Implement exclude patterns from workspace settings (#8037)"
This reverts commit 5fb5d0e9b9d128883195b2aeaa0aeea0052a5886.
