--- .sourcegraph/.ignore
@@ -1,1 +1,0 @@
-recordings/
