+++ lib/shared/package.json
@@ -35,7 +35,6 @@
     "lexical": "^0.17.0",
     "lodash": "^4.17.21",
     "lru-cache": "^10.0.0",
-    "minimatch": "^9.0.3",
     "ollama": "^0.5.1",
     "re2js": "^0.4.1",
     "semver": "^7.5.4",
